+++ src/models/attaques/BouletDeCanon.java
@@ -56,7 +56,7 @@ public class BouletDeCanon extends Attaque
      * @param cible la creature visee
      */
     public BouletDeCanon(Terrain terrain, Tour attaquant, Creature cible, 
-                        int degats, double rayonImpact)
+                        long degats, double rayonImpact)
     {
         super((int) attaquant.getCenterX(),(int) attaquant.getCenterY(), terrain, attaquant, cible);
